Fix squash on non-numeric columns (#80)
* Fix squash on non-numeric columns

When applying a squash, create a dict of how to aggregate all columns, so that
non-numeric columns are maintained. Fixes #76 and #23.

* Fix subgraph_sum to handle multi-index dataframes

When grabbing the values in ``columns`` for that row's node and all of its
decendents in a multi-indexed dataframe, this function was aggregating over
``node`` instead of ``node`` and ``rank`` (for caliper reader).