Remove unnecessary setting of _hatchet_nid in dataframe (#109)
Any reader or operation that modifies the graph was traversing the graph, then
traversing the node column of the dataframe to set the _hatchet_nid. Traversing
the dataframe is unnecessary, since updated node objects will be reflected in
both the graph and dataframe. This does not change the functionality, it
just removes an extra (duplicated) step that has been covered by traversing the
graph.