add missing rows for math operations on graphframes, fixes unify
This PR fixes unify in the case where there are some nodes that exist in x and
not y, and other nodes that exist in y and not x. The unified graph will
contain more nodes than exist in x or y. The resulting dataframe needs to have
the missing nodes.

assuming z = x + y:
- add helper function to append missing rows in y, but not x
- this way, the x's dataframe will contain the same number of nodes as in the
  unified graph
- give these appended rows a value of 0 for the inc/exc metrics
- when performing the add/sub/div, z will perform the operation correctly if
  the value is in x and y and vice versa