fix pandas warning in unify's missing nodes
Pandas was throwing a SettingwithCopyWarning in the _insert_missing_rows API.
This API adds a new column to the dataframe, and this warning lets us know that
our code may have modified the original dataframe, when we want it to have
modified a subset dataframe (not a view).

This PR resolves this by performing a join of the original dataframe with a
dataframe containing the new column (as opposed to appending the new column
directly to the original dataframe).