do not update list of inclusive metrics when no exclusive metrics exist
The list of exclusive metrics was being emptied in update_inclusive_columns()
when there was no corresponding exclusive metrics defined on the graphframe.
This fix will leave the lists of exclusive and inclusive metrics untouched if
there are no exclusive metrics defined. This does not impact the inclusive
metric values.