caliper-reader: modify missing data logic
If there is at least a row (i.e., rank) present for a given node (and other
ranks missing), then use existing values from present rank to fill data for
missing ranks. Previously, the current implementation is to fill all missing
data with None or 0 values, which caused a mismatch in data for the same node.