caliper native reader: refactor reader, recursively create parent nodes
There is no ordering scheme for caliper records, so when parsing the records to
build the tree, we may encounter a record for the child before we see a record
for its parent. This fix recursively builds the tree nodes for each record,
creating the child node as well as any of its parents.

Other changes:
- check for value columns when creating list of exc and inc metrics (check
  is_value() attribute for given column)
- change column names from non-aggregated cali file to time and time (inc) if
  present
- check for columns with (inc) or inclusive when creating list of inclusive
  metrics
- fix mismatch of column names in dataframe and list of inc and exc metrics